Remove all TODO tasks
